+++ lib/peddler/flat_file_parser.rb
@@ -1,10 +1,13 @@
 require 'delegate'
 require 'csv'
 require 'digest/md5'
+require 'peddler/headers'
 
 module Peddler
   # @api private
   class FlatFileParser < SimpleDelegator
+    include Headers
+
     # http://stackoverflow.com/questions/8073920/importing-csv-quoting-error-is-driving-me-nuts
     OPTIONS = { col_sep: "\t", quote_char: "\x00", headers: true }.freeze