+++ lib/peddler/xml_parser.rb
@@ -1,12 +1,14 @@
 require 'delegate'
 require 'dig_rb'
 require 'forwardable'
+require 'peddler/headers'
 require 'multi_xml'
 
 module Peddler
   # @api private
   class XMLParser < SimpleDelegator
     extend Forwardable
+    include Headers
 
     def_delegator :parse, :dig